fix: remove constants from types package, fix tsconfig
- Remove constants.ts from @kbn/data-views-types (constants don't belong in types package)
- Restore constants to data_views plugin
- Define RuntimeType as literal string union instead of derived from constant
- Fix tsconfig.json to use @kbn/tsconfig-base and proper kbn_references
- Add @kbn/data-views-types to data_views plugin's kbn_references
