+++ src/platform/plugins/shared/data_views/tsconfig.json
@@ -14,6 +14,7 @@
   ],
   "kbn_references": [
     "@kbn/core",
+    "@kbn/data-views-types",
     "@kbn/data-view-validation",
     "@kbn/usage-collection-plugin",
     "@kbn/kibana-utils-plugin",
